% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_fission.R
\name{chain_fission}
\alias{chain_fission}
\title{Perform post-clustering inference using chain data fission}
\usage{
chain_fission(
  X,
  Sigma = NULL,
  tau = 0.4,
  nFission = 1000,
  merge_function = c("KS", "Fisher", "geometric", "harmonic"),
  cl_fun,
  k,
  k1,
  k2,
  parallel = FALSE,
  ncores = NULL,
  test,
  ...
)
}
\arguments{
\item{X}{The data matrix of size n x p where post-clustering inference must be applied}

\item{Sigma}{An estimator of the covariance matrix of size p x p of X. If NULL, sample covariance estimator is used}

\item{tau}{The tuning parameter used for data fission}

\item{nFission}{The number of repeated data fissions to perform}

\item{merge_function}{The merging function used to derive chain fission p-values. It can be one of "KS", "Fisher", "geometric", or "harmonic"}

\item{cl_fun}{A clustering function that returns a factor vector containing the clusters}

\item{k}{The number of clusters to be built using cl_fun}

\item{k1}{The first cluster of interest among the k}

\item{k2}{The second cluster of interest among the k}

\item{parallel}{A logical flag indicating whether parallel computation should be enabled}

\item{ncores}{An integer indicating the number of cores to be used when parallel is TRUE}

\item{test}{A test function that inputs a numeric matrix, a partition and the two clusters to test and that outputs p-values}

\item{...}{Additional arguments that can be passed to the test function}
}
\value{
A list with the following elements:
\itemize{
  \item \code{Cluster}: The nFission estimated partitions
  \item \code{p.value}: A numeric vector of length p containing the nFission merging p-values or a list of M numeric vectors of length p if more than one merging method is used
}
}
\description{
Perform post-clustering inference using chain data fission
}
\examples{
cl_fun <- function(x, K) {
  km <- kmeans(x, center = K)
  return(km$cluster)
}

X1_cl1 <- matrix(rnorm(100 * 25, mean = 0), ncol = 25)
X1_cl2 <- matrix(rnorm(100 * 25, mean = 1.5), ncol = 25)
X1 <- rbind(X1_cl1, X1_cl2)
X2 <- matrix(rnorm(200 * 50), ncol = 50)
X <- cbind(X1, X2)
plot(X, col = rep(1:2, each = 100))
chain_fission(X, nFission = 10, cl_fun = cl_fun, k = 2, k1 = 1, k2 = 2, test = t_test.fission)
fission(X, tau = 0.4, cl_fun = cl_fun, k = 2, k1 = 1, k2 = 2, test = t_test.fission)

}
