% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fission.R
\name{data_fission}
\alias{data_fission}
\title{Data fission of Gaussian data}
\usage{
data_fission(X, Sigma = NULL, tau = 0.4)
}
\arguments{
\item{X}{The data matrix of size n x p for fission.}

\item{Sigma}{An estimator of the covariance matrix of size p x p  of X.}

\item{tau}{The tuning parameter used of data fission}
}
\value{
A list with the following elements \itemize{
\item \code{fX}: The dataset that will be used for clustering
\item \code{gX}: The dataset that will be used for inference}
}
\description{
Data fission of Gaussian data
}
\examples{
X <- matrix(rnorm(200), ncol = 2)
fission <- data_fission(X, tau = 0.4)
cl <- kmeans(fission$fX, centers = 2)
par(mfrow = c(1,3))
plot(X)
plot(fission$fX, col = cl$cluster)
plot(fission$gX, col = cl$cluster)
}
