% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dear_seq.fission.R
\name{dear_seq.fission}
\alias{dear_seq.fission}
\title{Perform Differential Expression Analysis usgin dearseq.}
\usage{
dear_seq.fission(X, cl, ...)
}
\arguments{
\item{X}{A matrix of size n x p containing the expression data.}

\item{cl}{A numeric design matrix of size n x K containing the K variables to be tested.}

\item{...}{Additional arguments to be passed to the \code{dear_seq} function.}
}
\value{
A numeric vector containing the p-values obtained from the dearseq analysis.
}
\description{
Perform Differential Expression Analysis usgin dearseq.
}
\examples{
# Generate example data
set.seed(123)
X <- matrix(rnorm(100), nrow = 10)
cl <- c(rep(1, 5), rep(2, 5))

# Perform DEAR-Seq analysis using dear_seq.fission
p_values <- dear_seq.fission(X, cl, preprocessed = T)

# Display the results
p_values

}
\references{
For more information on the dear_seq function, refer to its documentation: ?dear_seq
}
