% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dearseq_chain_fission.R
\name{dearseq_chain_fission}
\alias{dearseq_chain_fission}
\title{Post clustering inference using chain data fission and dearseq}
\usage{
dearseq_chain_fission(
  X,
  Sigma = NULL,
  tau = 0.4,
  nFission = 1000,
  preprocessed = FALSE,
  merge_function = c("KS", "Fisher", "geometric", "harmonic"),
  cl_fun,
  k,
  k1,
  k2,
  parallel = FALSE,
  ncores = NULL
)
}
\arguments{
\item{X}{The data matrix of size n x p where post-clustering inference must be applied}

\item{Sigma}{An estimator of the covariance matrix of size p x p  of X}

\item{tau}{The tuning parameter used of data fission}

\item{nFission}{The number of repeated data fission to perform}

\item{preprocessed}{A logical flag indicating whether the expression data
have already been preprocessed (e.g. log2 transformed).
Default is FALSE, in which case y is assumed to contain raw counts and is
normalized into log(counts) per million.}

\item{merge_function}{The merging function used to derive chain fission p values}

\item{cl_fun}{The clustering function that will applied on f(X)}

\item{k}{The number of clusters built using cl_fun}

\item{k1}{The first cluster of interest among the k}

\item{k2}{The second cluster of interest among the k}

\item{parallel}{A logical flag indicating whether parallel computation should be enabled}

\item{ncores}{An integer indicating the number of cores to be used when parallel is TRUE}
}
\value{
#' A list with the following elements \itemize{
\item \code{Cluster}: the nFission estimated partition on f(X)
\item \code{p.value}: a numeric vector of length p containing the nFission merging p-values or a list
of M numeric vectors of length p if more than one merging method are used}
}
\description{
Post clustering inference using chain data fission and dearseq
}
\examples{
cl_fun <- function(x, K){
km <- kmeans(x, center  = K)
  return(km$cluster)
}
X1 <- c(rnorm(100, sd = 1), rnorm(100, mean = 2, sd = 1))
X2 <- c(rnorm(200))
X <- cbind(X1,X2)
plot(X, col = rep(1:2, each = 100))
dearseq_chain_fission(X, nFission = 10, cl_fun = cl_fun, k=2,  k1=1, k2=2, preprocessed = TRUE)
dearseq_fission(X, tau = 0.4, cl_fun = cl_fun, k=2, k1=1, k2=2, preprocessed = TRUE)
}
