% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.test_fission.R
\name{t_test.fission}
\alias{t_test.fission}
\title{Perform independent two-sample t-tests on columns of a matrix or data frame.}
\usage{
t_test.fission(X, cl, ...)
}
\arguments{
\item{X}{A matrix of size n x p containing the data.}

\item{cl}{A vector of length n specifying the class labels for each observation in X.}

\item{...}{Additional arguments to be passed to the \code{t.test} function.}
}
\value{
A numeric vector of length p containing the p-values obtained from the t-tests.
}
\description{
Perform independent two-sample t-tests on columns of a matrix or data frame.
}
\examples{
# Generate example data
set.seed(123)
X <- matrix(rnorm(100), nrow = 10)
cl <- c(rep(1, 5), rep(2, 5))

# Perform t-test using t_test.fission
p_values <- t_test.fission(X, cl)

# Display the results
p_values

}
\references{
For more information on the t.test function, refer to its documentation: ?t.test
}
