% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_multiclusters.R
\name{t_test_multiclusters}
\alias{t_test_multiclusters}
\title{Performs t-test for K >= 2 clusters}
\usage{
t_test_multiclusters(x, cl, ...)
}
\arguments{
\item{x}{a numerical vectors of size n}

\item{cl}{A vector of length n specifying the class labels for each observation in x}

\item{...}{Additional arguments to be passed to the \code{t.test} function.}
}
\value{
A dataframe containing p-values for the t-test between each clusters pairs
}
\description{
Performs t-test for K >= 2 clusters
}
\examples{
n <- 150
cl <- sample(1:5, size = n, replace = TRUE)
x <- rnorm(n)
t_test_multiclusters(x, cl)
}
